+++ 대니-김빈/1주차/week1/1week/package.json
@@ -1,15 +1,13 @@
 {
-  "name": "week01",
+  "name": "1week",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "draft-js": "^0.11.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-draft-wysiwyg": "^1.15.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
