+++ 대니-김빈/1주차/week1/1week/public/manifest.json
