+++ 대니-김빈/1주차/week1/1week/src/EditBox.css
@@ -1,0 +1,28 @@
+/* EditBox.css */
+.edit-box-container {
+    display: flex;
+    flex-direction: column; /* 세로로 배치 */
+    align-items: center; /* 가로 방향으로 중앙 정렬 */
+    text-align: center;
+  }
+  
+  .edit-box {
+    width: 800px;
+    height: 40px;
+    font-size: 18px;  
+    padding: 8px;
+    background-color: #d1e4f2;
+    margin-bottom: 20px; /* 입력 상자와 할 일 목록 사이에 간격 추가 */
+  }
+  
+  ul {
+    list-style-type: none; /* 기본 불릿 제거 */
+    padding: 0;
+  }
+  
+  li {
+    font-size: 18px;
+    color: #333;
+    margin-bottom: 10px;
+  }
+
