+++ 대니-김빈/1주차/week1/1week/src/TaskList.css
@@ -10,13 +10,15 @@
     font-size: 18px;
     color: #333;
     display: flex;
+    padding: 5px 10px;
     justify-content: space-between; /* 할 일과 버튼 간의 간격 */
   }
   
   .register-button {
-    background-color: #bbe2fd;
+    background-color: #a1d1f4;
     color: black;
     border: none;
+    margin-left: 10px;
     padding: 5px 10px;
     cursor: pointer;
   }
